+++ tests/multipleMakerOrders.ts
@@ -14,36 +14,30 @@ import {
 	isVariant,
 	OracleSource,
 	PEG_PRECISION,
-	BulkAccountLoader,
 } from '../sdk/src';
 
 import {
 	createUserWithUSDCAccount,
 	initializeQuoteSpotMarket,
-	mockOracle,
+	mockOracleNoProgram,
 	mockUSDCMint,
 	mockUserUSDCAccount,
-	printTxLogs,
-	setFeedPrice,
+	setFeedPriceNoProgram,
 } from './testHelpers';
 import { ContractTier, MARGIN_PRECISION, OrderType } from '../sdk';
+import { startAnchor } from 'solana-bankrun';
+import { TestBulkAccountLoader } from '../sdk/src/accounts/testBulkAccountLoader';
+import { BankrunContextWrapper } from '../sdk/src/bankrun/bankrunConnection';
 
 describe('multiple maker orders', () => {
-	const provider = anchor.AnchorProvider.local(undefined, {
-		commitment: 'confirmed',
-		preflightCommitment: 'confirmed',
-	});
-	const connection = provider.connection;
-	anchor.setProvider(provider);
 	const chProgram = anchor.workspace.Drift as Program;
 
 	let fillerDriftClient: TestClient;
-	const eventSubscriber = new EventSubscriber(connection, chProgram, {
-		commitment: 'recent',
-	});
-	eventSubscriber.subscribe();
+	let eventSubscriber: EventSubscriber;
 
-	const bulkAccountLoader = new BulkAccountLoader(connection, 'confirmed', 1);
+	let bulkAccountLoader: TestBulkAccountLoader;
+
+	let bankrunContextWrapper: BankrunContextWrapper;
 
 	let usdcMint;
 	let userUSDCAccount;
@@ -66,19 +60,40 @@ describe('multiple maker orders', () => {
 	let oracleInfos;
 
 	before(async () => {
-		usdcMint = await mockUSDCMint(provider);
-		userUSDCAccount = await mockUserUSDCAccount(usdcMint, usdcAmount, provider);
+		const context = await startAnchor('', [], []);
 
-		solUsd = await mockOracle(100);
-		dogUsd = await mockOracle(0.6899, -4, 0);
+		bankrunContextWrapper = new BankrunContextWrapper(context);
+
+		bulkAccountLoader = new TestBulkAccountLoader(
+			bankrunContextWrapper.connection,
+			'processed',
+			1
+		);
+
+		eventSubscriber = new EventSubscriber(
+			bankrunContextWrapper.connection.toConnection(),
+			chProgram
+		);
+
+		await eventSubscriber.subscribe();
+
+		usdcMint = await mockUSDCMint(bankrunContextWrapper);
+		userUSDCAccount = await mockUserUSDCAccount(
+			usdcMint,
+			usdcAmount,
+			bankrunContextWrapper
+		);
+
+		solUsd = await mockOracleNoProgram(bankrunContextWrapper, 100);
+		dogUsd = await mockOracleNoProgram(bankrunContextWrapper, 0.6899, -4, 0);
 
 		marketIndexes = [0, 1];
 		spotMarketIndexes = [0];
 		oracleInfos = [{ publicKey: solUsd, source: OracleSource.PYTH }];
 
 		fillerDriftClient = new TestClient({
-			connection,
-			wallet: provider.wallet,
+			connection: bankrunContextWrapper.connection.toConnection(),
+			wallet: bankrunContextWrapper.provider.wallet,
 			programID: chProgram.programId,
 			opts: {
 				commitment: 'confirmed',
@@ -86,6 +101,7 @@ describe('multiple maker orders', () => {
 			activeSubAccountId: 0,
 			perpMarketIndexes: marketIndexes,
 			spotMarketIndexes: spotMarketIndexes,
+			subAccountIds: [],
 			oracleInfos,
 			accountSubscription: {
 				type: 'polling',
@@ -166,7 +182,7 @@ describe('multiple maker orders', () => {
 	it('taker long solUsd', async () => {
 		const [takerDriftClient, takerUSDCAccount] =
 			await createUserWithUSDCAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				usdcAmount,
@@ -180,7 +196,7 @@ describe('multiple maker orders', () => {
 
 		const [makerDriftClient, makerUSDCAccount] =
 			await createUserWithUSDCAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				usdcAmount,
@@ -204,7 +220,7 @@ describe('multiple maker orders', () => {
 
 		const [secondMakerDriftClient, secondMakerUSDCAccount] =
 			await createUserWithUSDCAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				usdcAmount,
@@ -254,7 +270,7 @@ describe('multiple maker orders', () => {
 			makerInfo
 		);
 
-		await printTxLogs(connection, txSig);
+		bankrunContextWrapper.printTxLogs(txSig);
 
 		const orderActionRecords = eventSubscriber
 			.getEventsArray('OrderActionRecord')
@@ -303,7 +319,7 @@ describe('multiple maker orders', () => {
 			});
 		}
 
-		await setFeedPrice(anchor.workspace.Pyth, 90, solUsd);
+		await setFeedPriceNoProgram(bankrunContextWrapper, 90, solUsd);
 		await takerDriftClient.placePerpOrder({
 			marketIndex: 0,
 			orderType: OrderType.LIMIT,
@@ -322,7 +338,7 @@ describe('multiple maker orders', () => {
 		const takerPosition2 = takerDriftClient.getUser().getPerpPosition(0);
 		assert(takerPosition2.baseAssetAmount.eq(new BN(0)));
 
-		await printTxLogs(connection, txSig2);
+		bankrunContextWrapper.printTxLogs(txSig2);
 
 		await takerDriftClient.unsubscribe();
 		await makerDriftClient.unsubscribe();
@@ -332,7 +348,7 @@ describe('multiple maker orders', () => {
 	it('taker short dogUsd', async () => {
 		const [takerDriftClient, takerUSDCAccount] =
 			await createUserWithUSDCAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				usdcAmount,
@@ -346,7 +362,7 @@ describe('multiple maker orders', () => {
 
 		const [makerDriftClient, makerUSDCAccount] =
 			await createUserWithUSDCAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				usdcAmount,
@@ -370,7 +386,7 @@ describe('multiple maker orders', () => {
 
 		const [secondMakerDriftClient, secondMakerUSDCAccount] =
 			await createUserWithUSDCAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				usdcAmount,
@@ -394,7 +410,7 @@ describe('multiple maker orders', () => {
 
 		const [thirdMakerDriftClient, thirdMakerUSDCAccount] =
 			await createUserWithUSDCAccount(
-				provider,
+				bankrunContextWrapper,
 				usdcMint,
 				chProgram,
 				usdcAmount,
@@ -425,7 +441,7 @@ describe('multiple maker orders', () => {
 			}
 		}
 
-		await setFeedPrice(anchor.workspace.Pyth, 0.675, dogUsd);
+		await setFeedPriceNoProgram(bankrunContextWrapper, 0.675, dogUsd);
 		const takerBaseAssetAmount = new BN(600).mul(BASE_PRECISION);
 		await takerDriftClient.placePerpOrder({
 			marketIndex: 1,
@@ -458,8 +474,7 @@ describe('multiple maker orders', () => {
 			takerDriftClient.getOrder(1),
 			makerInfo
 		);
-
-		await printTxLogs(connection, txSig);
+		bankrunContextWrapper.printTxLogs(txSig);
 
 		const orderActionRecords = eventSubscriber
 			.getEventsArray('OrderActionRecord')
@@ -548,7 +563,7 @@ describe('multiple maker orders', () => {
 			});
 		}
 
-		await setFeedPrice(anchor.workspace.Pyth, 0.75, dogUsd);
+		await setFeedPriceNoProgram(bankrunContextWrapper, 0.75, dogUsd);
 		await takerDriftClient.placePerpOrder({
 			marketIndex: 1,
 			orderType: OrderType.LIMIT,
@@ -580,7 +595,7 @@ describe('multiple maker orders', () => {
 			dogMarketAfter.amm.baseAssetAmountWithAmm.eq(new BN('-66279600000'))
 		);
 
-		await printTxLogs(connection, txSig2);
+		bankrunContextWrapper.printTxLogs(txSig2);
 
 		await takerDriftClient.unsubscribe();
 		await makerDriftClient.unsubscribe();